Updated Auth functionality to handle page refresh
