+++ src/backend/db/users.js
@@ -13,6 +13,10 @@ export const users = [
     lastName: "Balika",
     email: "adarshbalika@gmail.com",
     password: "adarshBalika@123",
+    likes: [],
+    history: [],
+    playlists: [],
+    watchlater: [],
     createdAt: formatDate(),
     updatedAt: formatDate(),
   }
